review: break github-app↔github-api cycle, redirect:manual on members probe (P2)
Multi-agent review (security-sentinel, architecture-strategist,
code-quality-analyst, user-impact-reviewer, semgrep) — no P1. Inline fixes:

- P2 (security): probeOrgMembership's githubFetch inherited redirect:"follow",
  so the 302 "requester not visible as org member" branch was unreachable in
  prod (GitHub auto-redirected to /public_members). Add redirect:"manual" to
  surface 302 as an authoritative not-member, matching the verifyInstallationOwnership
  precedent. Fail-closed either way (review confirmed no escalation), but this
  makes the documented 302 branch reachable + the test honest.
- P2 (architecture + code-quality): github-app.ts importing isRetryable from
  github-api.ts (which imports generateInstallationToken from github-app.ts)
  closed an import cycle, and delay was duplicated. Extract both into a
  dependency-free github-retry.ts leaf both modules point down to.
- P3 (code-quality): real-factory default mock mockResolvedValue(null) was
  shape-stale vs the new { installationId, outcome } return.

Wontfix (noted): ownerInstall===installationId fall-through (unreachable);
no-github-login variant (tested public-boundary guard, not dead).

tsc clean; full web-platform vitest suite green (8992 passed).
